fix: soft restart camera

+++ base/wolfvision.cpp
@@ -5,8 +5,7 @@ int main() {
   fmt::print("[{}] WolfVision config file path: {}\n", idntifier, CONFIG_FILE_PATH);
 
   cv::Mat src_img_, roi_img_;
-
-  mindvision::VideoCapture mv_capture_ = mindvision::VideoCapture(
+  mindvision::VideoCapture* mv_capture_ = new mindvision::VideoCapture(
     mindvision::CameraParam(0, mindvision::RESOLUTION_1280_X_800, mindvision::EXPOSURE_600));
 
   uart::SerialPort serial_ = uart::SerialPort(
@@ -37,8 +36,8 @@ int main() {
   basic_roi::RoI roi_;
   while (true) {
     global_fps_.getTick();
-    if (mv_capture_.isindustryimgInput()) {
-      src_img_ = mv_capture_.image();
+    if (mv_capture_->isindustryimgInput()) {
+      src_img_ = mv_capture_->image();
     } else {
       cap_.read(src_img_);
     }
@@ -151,7 +150,7 @@ int main() {
     if (serial_.returnReceiveMode() != uart::SENTRY_STRIKE_MODE) {
       basic_armor_.initializationSentryMode();
     }
-    mv_capture_.cameraReleasebuff();
+    mv_capture_->cameraReleasebuff();
     basic_armor_.freeMemory();
 
 #ifndef RELEASE
@@ -165,7 +164,18 @@ int main() {
 #endif
     global_fps_.calculateFPSGlobal();
     if (global_fps_.returnFps() > 500) {
-      int i[[maybe_unused]] = std::system("echo 1 | sudo -S reboot");
+      mv_capture_->~VideoCapture();
+      while (!utils::resetMVCamera()) {
+        usleep(100);
+      }
+      usleep(100);
+      mv_capture_ = new mindvision::VideoCapture(mindvision::CameraParam(
+          0, mindvision::RESOLUTION_1280_X_800, mindvision::EXPOSURE_600));
+
+      static int counter { 30 };
+      if (!--counter) {
+        int i [[maybe_unused]] = std::system("echo 1 | sudo -S reboot");
+      }
     }
   }
   return 0;
